fixes issue 589 Add support for PECOFF + DWARF symbols on Windows
Major code refactoring:

  drsyms_linux.c --> drsyms_unix.c --> drsyms_dwarf.c
                       ^    |  |
                       |    |  +--->   drsyms_elf.c
  drsyms_windows.c ----+    +------>   drsyms_pecoff.c
              |
              +---------> dbghelp.dll

+ drsyms_windows.c asks drsyms_unix.c to load a module upon first
  occurence and see whether it has pecoff symbols: if so, all
  further routines are routed to drsyms_unix.c
+ drsyms_unix.c talks to drsyms_{elf,pecoff}.c via drsym_obj_ API

Changes to existing drsyms code:
+ b/c cygwin/mingw debug info has no lowpc/highpc in CU die,
  added use of .debug_aranges to locate CU for given address
  before using current method

Added drsyms_pecoff.c:
+ uses pecoff symbol and string tables for symbol lookup, mirroring
  ELF, since these sections don't seem to be stripped by default
+ uses dwarf via new refactoring to get line# info
+ there is no symbol end pc (missing aux entry for symbols)
  so uses next symbol as end pc
+ when sorting symbols, take larger Type value when have two identical
  offset symbol entries
+ remove extra leading underscores when present (try to detect)

Docs and tests:
+ update documentation on windows pdb vs pecoff
+ added test: build drsyms-test and its lib w/ gcc
  if gcc is found on the system.
  had to change the test a bunch to get this to work
  b/c it had hardcoded names.
  also fixed a bug in the test where it demangled
  every other test string.
+ no 64-bit test b/c no 64-bit cygwin
  XXX: get x64 MinGW

Added libelftc build for Windows PECOFF:
+ added libelftc-pecoff binary libraries and symbols for 32-bit and
  64-bit (64-bit was manually tested).
+ added source patch and instructions for building dwarf.lib and elftc.lib
+ binaries are built with VS2005 since VS2008-built break VS2005 link
+ changed libelftc to store truncated mangled name and return
  size needed for pecoff

Other:
+ fixed bug in DR on COW applied to read-only mapped file

SVN-Revision: 1250

+++ ext/drsyms/demangle.cc
@@ -212,7 +212,7 @@ static void InitState(State *state, const char *mangled,
 
 // Calculates the remaining length of the mangled name.
 static int RemainingLength(State *state) {
-  return state->mangled_end - state->mangled_cur;
+  return (int) (state->mangled_end - state->mangled_cur);
 }
 
 // Returns true and advances "mangled_cur" if we find "c" at
@@ -378,7 +378,7 @@ static void MaybeAppendWithLength(State *state, const char * const str,
 // A convenient wrapper arount MaybeAppendWithLength().
 static bool MaybeAppend(State *state, const char * const str) {
   if (state->append) {
-    int length = StrLen(str);
+    int length = (int) StrLen(str);
     MaybeAppendWithLength(state, str, length);
   }
   return true;
